Revert "refactor: made the Jinja2Template extends the ATemplate class, allowing the future documentation generation method to be tested in isolation IHS-201"
This reverts commit 03fa4027d11cf34704351dcbf92330bfe908be42.
